% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{getRandomNum}
\alias{getRandomNum}
\title{Random number generation based on order statistics}
\usage{
getRandomNum(x, min, q25, q50, q75, max)
}
\arguments{
\item{x}{Sample size}

\item{min}{minimum}

\item{q25}{quantile}

\item{q50}{quantile}

\item{q75}{quantile}

\item{max}{maximum}
}
\value{
A Random sequence
}
\description{
Random number generation based on order statistics
}
\examples{
\dontrun{
z = getRandomNum(x = 5,min = 2230,q25 = 18123,q50 = 52213,q75 = 78312,max = 234234123)
quantile(z)
fivenum(z)
}
}
