% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{GradientDescent}
\alias{GradientDescent}
\title{Gradient descent method}
\usage{
GradientDescent(x, y, error, maxiter, stepmethod, step, alpha, beta)
}
\arguments{
\item{x}{X is the data matrix}

\item{y}{Y  is measurements}

\item{error}{Error termination condition}

\item{maxiter}{Maximum number of iterations for Maxiter}

\item{stepmethod}{The default is Backward descent method}

\item{step}{Step is the fixed step size set}

\item{alpha}{It is a parameter of the backward descent method}

\item{beta}{It is a parameter of the backward descent method}
}
\value{
Coefficient,iteration number,error
}
\description{
Gradient descent method
}
\examples{
\dontrun{
data("iris")
x<-matrix(iris[1:50,1],50,1)
y<-matrix(iris[1:50,2],50,1)
 GradientDescent(x,y,1e-14,1000,stepmethod=T,step=0.001,alpha=0.25,beta=0.8)
}
}
