% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{irisl}
\alias{irisl}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of Gradient Descent.
}
\examples{
\dontrun{
data(irisl)
x<-matrix(iris[1:50,1],50,1)
y<-matrix(iris[1:50,2],50,1)
GradientDescent(x,y,1e-14,1000,stepmethod=T,step=0.001,alpha=0.25,beta=0.8)
}
}
