% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mean_ci}
\alias{mean_ci}
\title{The confidence interval calculate}
\usage{
mean_ci(x, conf)
}
\arguments{
\item{x}{a input list confidence}

\item{conf}{a parameter of confidence}
}
\value{
A confidence interval
}
\description{
The confidence interval between the two ends of the mean is calculated using an approximate normal distribution
}
\examples{
\dontrun{
conf=0.95
x <- runif(100)
mean_ci(x,conf)
mean_ci(x, conf = 0.99)
}
}
