% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisC}
\alias{MetropolisC}
\title{A  Metropolis samplered using Rcpp}
\usage{
MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{the sigma of samples}

\item{x0}{a parameter}

\item{N}{sample of size}
}
\value{
a random sample of size \code{N}
}
\description{
A Metropolis samplered using Rcpp
}
\examples{
\dontrun{
x0<-25
sigma<-c(0.05,1,2,16)
N<-3000
M1=MetropolisC(sigma[1],x0,N)
M2=MetropolisC(sigma[2],x0,N)
M3=MetropolisC(sigma[3],x0,N)
M4=MetropolisC(sigma[4],x0,N)
acceptanceC<-1-c(M1[[2]][1],M2[[2]][1],M3[[2]][1],M4[[2]][1])/N
index<-1:2000
My1<-M1[[1]][index]
My2<-M2[[1]][index]
My3<-M3[[1]][index]
My4<-M4[[1]][index]
opar<-par(no.readonly = T)
par(mfrow=c(2,2))
plot(index,My1,type = "l",xlab = "sigma=0.05")
plot(index,My2,type = "l",xlab = "sigma=1")
plot(index,My3,type = "l",xlab = "sigma=2")
plot(index,My4,type = "l",xlab = "sigam=16")
}
}
