% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{myst}
\alias{myst}
\title{Basic statistic}
\usage{
myst(x)
}
\arguments{
\item{x}{a input list}
}
\value{
A list about some basic statistics
}
\description{
Mean, variance, sample size, kurtosis, skewness
}
\examples{
\dontrun{
x=c(1,2,3,4)
mystat(x)
}
}
