\name{StatComp20058-package}
\alias{StatComp20058-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package ( \code{MetropolisR} \code{gibbsR} and \code{vaccR}  versus \code{MetropolisC} \code{gibbsC} and \code{vaccC})
  The R package also includes basic statistical analysis \code{myst} , approximate confidence interval estimation \code{mean_ci} , random sequence generation based on order statistics \code{getRandomNum}, and detection of outliers \code{outlierKD} by four functions.
}
\author{
Xinyu Li

Maintainer: Xinyu Li <xinyuli188@163.com>
}
\keyword{ package }
