\name{StatComp20058-package}
\alias{StatComp20058-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} versus \code{gibbsC} and \code{vaccC})
}
\author{
Xinyu Li

Maintainer: Xinyu Li <xinyuli188@163.com>
}
\keyword{ package }
